requirements.txt
absl-py==0.12.0
argcomplete==1.12.3
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
colorama==0.4.4
colorlog==6.5.0
contextlib2==21.6.0
cycler==0.11.0
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
devfac-py-logger==0.2.2
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
entrypoints==0.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin==0.1.6
gin-config==0.5.0
google-api-core==2.2.1
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
h5py==3.1.0
httplib2==0.19.1
idna==3.3
imgviz==1.4.1
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.0
jupyterlab-pygments==0.1.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labelme==4.5.13
lvis==0.5.3
lxml==4.6.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.2.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.19.5
oauthlib==3.1.1
object-detection==0.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.2
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
portalocker==2.3.2
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools-windows==2.0.0.2
pydot==1.4.2
Pygments==2.10.0
pymongo==3.12.1
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pywin32==302
PyYAML==6.0
pyzmq==22.3.0
QtPy==1.11.2
regex==2021.11.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==2.0.0
scikit-learn==1.0.1
scipy==1.7.1
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
sklearn==0.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.4.0
tensorflow-model-optimization==0.7.0
tensorflow-text==2.6.0
termcolor==1.1.0
testpath==0.5.0
text-unidecode==1.3
tf-slim==1.1.0
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typeguard==2.13.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
